Add the LogReplicationSenderBuffer mechanism.    * Simplfy the LogEntrySender logic.    * Start siteManager while DiscoveryService is up.
