+++ log-replication/src/main/java/org/corfudb/logreplication/send/LogEntrySender.java
@@ -1,7 +1,5 @@
 package org.corfudb.logreplication.send;
 
-import lombok.Data;
-import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 import org.corfudb.infrastructure.logreplication.DataSender;
@@ -16,18 +14,11 @@ import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.view.Address;
 import org.corfudb.runtime.exceptions.TrimmedException;
 
-import java.io.File;
-import java.io.FileReader;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Properties;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import java.util.stream.Collectors;
 
 /**
  * This class is responsible of managing the transmission of log entries,
@@ -38,38 +29,8 @@ import java.util.stream.Collectors;
  */
 @Slf4j
 public class LogEntrySender {
-    public static final String config_file = "/config/corfu/corfu_replication_config.properties";
-    public static final int DEFAULT_READER_QUEUE_SIZE = 1;
-    public static final int DEFAULT_RESENT_TIMER = 5000;
-    public static final int DEFAULT_MAX_RETRY = 5;
+
     public static final int DEFAULT_TIMEOUT = 5000;
-
-    /*
-     * for internal timer increasing for each message in milliseconds
-     */
-    final static private long TIME_INCREMENT = 10;
-
-    private int readerBatchSize = DEFAULT_READER_QUEUE_SIZE;
-
-    /*
-     * The timer to resend an entry. This is the roundtrip time between sender/receiver.
-     */
-    private int msgTimer = DEFAULT_RESENT_TIMER;
-
-    /*
-     * The max number of retry for sending an entry.
-     */
-    private int maxRetry = DEFAULT_MAX_RETRY;
-
-    /*
-     * wait for ack or not
-     */
-    private boolean errorOnMsgTimeout = true;
-
-    /*
-     * reset while process messages
-     */
-    long currentTime;
 
     /*
      * Corfu Runtime
@@ -81,16 +42,12 @@ public class LogEntrySender {
      */
     private LogEntryReader logEntryReader;
 
-    /*
-     * Log Entry Listener, application callback to send out reads.
-     */
     private DataSender dataSender;
 
-    /*
-     * The log entry has been sent to the receiver but hasn't ACKed yet.
-     */
-    @Getter
-    LogReplicationEntryQueue pendingEntries;
+   /*
+    * Implementation of buffering messages and sending/resending messages
+    */
+    private LogReplicationSenderBuffer dataSenderBuffer;
 
     private long ackTs = Address.NON_ADDRESS;
 
@@ -102,8 +59,9 @@ public class LogEntrySender {
 
     private volatile boolean taskActive = false;
 
+    long currentTime;
 
-    private Map<Long, CompletableFuture<LogReplicationEntry>> pendingLogEntriesAcked = new HashMap<>();
+    //private Map<Long, CompletableFuture<LogReplicationEntry>> pendingLogEntriesAcked = new HashMap<>();
     private List<CompletableFuture<LogReplicationEntry>> pendingForAck = new ArrayList<>();
 
     /**
@@ -124,60 +82,13 @@ public class LogEntrySender {
     public LogEntrySender(CorfuRuntime runtime, LogEntryReader logEntryReader, DataSender dataSender,
                           ReadProcessor readProcessor, LogReplicationFSM logReplicationFSM) {
 
-        readConfig();
         this.runtime = runtime;
         this.logEntryReader = logEntryReader;
         this.dataSender = dataSender;
+        this.dataSenderBuffer = new LogReplicationSenderBuffer(dataSender);
         this.logReplicationFSM = logReplicationFSM;
-        this.pendingEntries = new LogReplicationEntryQueue(readerBatchSize);
+        //this.pendingEntries = new LogReplicationSenderQueue(readerBatchSize);
     }
-
-    private void readConfig() {
-        try {
-            File configFile = new File(config_file);
-            FileReader reader = new FileReader(configFile);
-
-            Properties props = new Properties();
-            props.load(reader);
-
-            maxRetry = Integer.parseInt(props.getProperty("log_reader_max_retry", Integer.toString(DEFAULT_MAX_RETRY)));
-            readerBatchSize = Integer.parseInt(props.getProperty("log_reader_queue_size", Integer.toString(DEFAULT_READER_QUEUE_SIZE)));
-            msgTimer = Integer.parseInt(props.getProperty("log_reader_resend_timer", Integer.toString(DEFAULT_RESENT_TIMER)));
-            errorOnMsgTimeout = Boolean.parseBoolean(props.getProperty("log_reader_error_on_message_timeout", "true"));
-            reader.close();
-
-        } catch (Exception e) {
-            log.warn("The config file is not available {} , will use the default vaules for config.", e.getCause());
-        } finally {
-            log.info("log logreader config max_retry {} reader_queue_size {} entry_resend_timer {} waitAck {}",
-                    maxRetry, readerBatchSize, msgTimer, errorOnMsgTimeout);
-        }
-    }
-
-    /**
-     * resend the messages in the queue if it times out.
-     * @param
-     * @return it returns false if there is an entry has been resent MAX_RETRY and timeout again.
-     * Otherwise it returns true.
-     */
-
-    void resend() {
-        for (int i = 0; i < pendingEntries.list.size() && taskActive; i++) {
-            LogReplicationPendingEntry entry  = pendingEntries.list.get(i);
-            if (entry.timeout(getCurrentTime(), msgTimer)) {
-                if (errorOnMsgTimeout && entry.retry >= maxRetry) {
-                    log.warn("Entry {} of type {} has been resent max times {} for timer {}.", entry.getData().getMetadata().getTimestamp(),
-                            entry.getData().getMetadata().getMessageMetadataType(), maxRetry, msgTimer);
-                    throw new LogEntrySyncTimeoutException("timeout");
-                }
-
-                entry.retry(getCurrentTime());
-                CompletableFuture<LogReplicationEntry> cf = dataSender.send(entry.getData());
-                pendingLogEntriesAcked.put(entry.getData().getMetadata().getTimestamp(), cf);
-                log.info("resend message " + entry.getData().getMetadata().getTimestamp());
-            }
-        }
-   }
 
     /**
      * Read and send incremental updates (log entries)
@@ -188,21 +99,9 @@ public class LogEntrySender {
 
         try {
             // If there are pending entries, resend them.
-            if (!pendingEntries.list.isEmpty()) {
+            if (!dataSenderBuffer.getPendingEntries().isEmpty()) {
                 try {
-                    LogReplicationEntry ack = (LogReplicationEntry)CompletableFuture.anyOf(pendingLogEntriesAcked
-                            .values().toArray(new CompletableFuture<?>[pendingLogEntriesAcked.size()])).get(DEFAULT_TIMEOUT, TimeUnit.MILLISECONDS);
-                    log.trace("Received Log Entry ack {}", ack.getMetadata());
-
-                    updateAckTs(ack.getMetadata().getTimestamp());
-
-                    // Remove all CFs for all entries with lower timestamps than that of the ACKed LogReplicationEntry
-                    // This is because receiver can send aggregated ACKs.
-                    pendingLogEntriesAcked = pendingLogEntriesAcked.entrySet().stream()
-                            .filter(entry -> entry.getKey() > ack.getMetadata().getTimestamp())
-                            .collect(Collectors.toMap(x -> x.getKey(), x -> x.getValue()));
-
-                    log.trace("Total pending log entry acks: {}, for timestamps: {}", pendingLogEntriesAcked.size(), pendingLogEntriesAcked.keySet());
+                    LogReplicationEntry ack = dataSenderBuffer.processAcks();
 
                     // Enforce a Log Entry Sync Replicated (ack) event, which will update metadata information
                     // Todo (Consider directly updating Corfu Metadata information here, without going through FSM)
@@ -210,14 +109,14 @@ public class LogEntrySender {
                                 new LogReplicationEventMetadata(ack.getMetadata().getSyncRequestId(), ack.getMetadata().getTimestamp())));
 
                 } catch (TimeoutException te) {
-                    log.info("Log Entry ACK timed out, pending acks for {}", pendingLogEntriesAcked.keySet());
+                    log.info("Log Entry ACK timed out, pending acks for {}", dataSenderBuffer.getPendingLogEntriesAcked().keySet());
                 } catch (Exception e) {
                     log.error("Exception caught while waiting on log entry ACK.", e);
                     cancelLogEntrySync(LogReplicationError.UNKNOWN, LogReplicationEventType.SYNC_CANCEL, logEntrySyncEventId);
                     return;
                 }
 
-                resend();
+                dataSenderBuffer.resend();
             }
         } catch (LogEntrySyncTimeoutException te) {
             log.error("LogEntrySyncTimeoutException after several retries.", te);
@@ -225,18 +124,14 @@ public class LogEntrySender {
             return;
         }
 
-        while (taskActive && pendingEntries.list.size() < readerBatchSize) {
+        while (taskActive && !dataSenderBuffer.getPendingEntries().isFull()) {
             LogReplicationEntry message;
             // Read and Send Log Entries
             try {
                 message = logEntryReader.read(logEntrySyncEventId);
 
                 if (message != null) {
-                    pendingEntries.append(message, getCurrentTime());
-                    CompletableFuture<LogReplicationEntry> cf = dataSender.send(message);
-                    log.debug("sending data %s", message.getMetadata());
-                    pendingLogEntriesAcked.put(message.getMetadata().getTimestamp(), cf);
-                    log.trace("send message " + message.getMetadata());
+                    dataSenderBuffer.sendWithBuffering(message);
                 } else {
                     // If no message is returned we can break out and enqueue a CONTINUE, so other processes can
                     // take over the shared thread pool of the state machine
@@ -275,6 +170,10 @@ public class LogEntrySender {
 
     }
 
+    public void updateAckTs(long ts) {
+        dataSenderBuffer.updateAckTs(ts);
+    }
+
     /**
      * Reset the log entry sender to initial state
      */
@@ -283,88 +182,6 @@ public class LogEntrySender {
         log.info("Reset baseSnapshot %s ackTs %s", ts0, ts1);
         logEntryReader.reset(ts0, ts1);
         ackTs = ts1;
-        pendingEntries.evictAll();
-    }
-
-    /**
-     * Update the last ackTimestamp and evict all entries whose timestamp is less or equal to the ackTimestamp
-     * @param ackTimestamp
-     */
-    public void updateAckTs(Long ackTimestamp) {
-        if (ackTimestamp <= ackTs)
-            return;
-        ackTs = ackTimestamp;
-        log.info("Pending entries before eviction at {} is {}", ackTs, pendingEntries.list.size());
-        pendingEntries.evictAll(ackTs);
-        log.info("Pending entries AFTER eviction at {} is {}", ackTs, pendingEntries.list.size());
-
-        log.trace("ackTS " + ackTs + " queue size " + pendingEntries.list.size());
-    }
-
-    private long getCurrentTime() {
-        currentTime += TIME_INCREMENT;
-        return currentTime;
-    }
-
-    /**
-     * The element kept in the sliding windown to remember the log entries sent over but hasn't been acknowledged by the
-     * receiver and we use the time to decide when a re-send is necessary.
-     */
-    @Data
-    public static class LogReplicationPendingEntry {
-        // The address of the log entry
-        org.corfudb.protocols.wireprotocol.logreplication.LogReplicationEntry data;
-
-        // The first time the log entry is sent over
-        long time;
-
-        // The number of retries for this entry
-        int retry;
-
-        public LogReplicationPendingEntry(org.corfudb.protocols.wireprotocol.logreplication.LogReplicationEntry data, long time) {
-            this.data = data;
-            this.time = time;
-            this.retry = 0;
-        }
-
-        boolean timeout(long ctime, long timer) {
-            log.trace("current time {} - original time {} = {} timer {}", ctime, this.time, timer);
-            return  (ctime - this.time) > timer;
-        }
-
-        void retry(long time) {
-            this.time = time;
-            retry++;
-        }
-    }
-
-    /**
-     * The sliding window to record the pending entries that have sent to the receiver but hasn't got an ACK yet.
-     * The alternative is to remember the address only and reset the stream head to rereading the data if the queue size
-     * is quite big.
-     */
-    public static class LogReplicationEntryQueue {
-
-        int size;
-        ArrayList<LogReplicationPendingEntry> list;
-
-        public LogReplicationEntryQueue(int size) {
-            this.size = size;
-            list = new ArrayList<>();
-        }
-
-        public void evictAll() {
-            list = new ArrayList<>();
-        }
-
-        void evictAll(long address) {
-            log.trace("evict address " + address);
-            list.removeIf(a->(a.data.getMetadata().getTimestamp() <= address));
-        }
-
-        void append(org.corfudb.protocols.wireprotocol.logreplication.LogReplicationEntry data, long timer) {
-            LogReplicationPendingEntry entry = new LogReplicationPendingEntry(data, timer);
-            list.add(entry);
-        }
+        dataSenderBuffer.getPendingEntries().evictAll();
     }
 }
